+++ test/smokes/rails_best_practices/exclude_multiple/foo.rb
@@ -1,5 +1,0 @@
-begin
-  foo
-rescue Exception => e
-  puts e.message
-end