+++ test/smokes/rails_best_practices/sandbox_rails/app/models/user.rb
@@ -1,4 +1,0 @@
-class User < ApplicationRecord
-  belongs_to :group
-  default_scope order('created_at desc')
-end