+++ test/smokes/rails_best_practices/sandbox_rails/db/migrate/20161216021234_add_group_id_to_users.rb
@@ -1,5 +1,0 @@
-class AddGroupIdToUsers < ActiveRecord::Migration[5.0]
-  def change
-    add_column :users, :group_id, :integer
-  end
-end