+++ test/smokes/rails_best_practices/valid_sideci_yml/a.rb
@@ -1,9 +1,0 @@
-class AController
-  def search
-    begin
-      foo.save
-    rescue Exception => e
-      logger.warn "Unable to foo, will ignore: #{e}"
-    end
-  end
-end